PKG -- [fcl] Adds prep template util for mutate and query
